Move to Bigarray
Instead of using an int array array we now use a Bigarray with
int8_unsigned values. This makes drawing more straightforward as each
entry corresponds to one pixel. Furthermore it significantly improves
the performance. The steps matrix -> rows -> list -> send are for
heavier than matrix -> send. This could perhaps be further optimised
with the Bigarray.Array2.sub_left function but as we are now at
negligible speedlosses this is not further investigated.

We are at 210-220 kB/s on SPI because I miscalculated the value in a
previous commit as each send corresponds to two bytes.
The buffer has 0.48MB and sending a full buffer now takes about 2.2s.
Using 4bpp or even 2bpp could push this far lower, perhaps even below
0.5s.