Implemented 1,2,4 bit per pixel
Implemented support for 1,2,4 bpp transfers.
As the IT8951 does not support 1bpp it is quite ugly, hopefully this can
be improved somewhat. There are no speed advantages to 3bpp so it has
not been implemented yet (or ever).
The code to get the correct bits out of the buffer is (get_16bit) is
currently overly complicated. Before modifying this code learning about
optimising OCaml is required.