[Buggy] Large amount of changes, added text module
This commit contains multiple somewhat WIP changes.
* Introduction of the Text module, still requires: string drawing,
blit of bigarrays (this requires changes to pcf-format)
* Confusion about x/y. Fixing the x/y axes in arrays is harder than
expected because the rotation does not work as expected. When changing
the rotation the width/height changes and therefore the buffer must be
recreated. Furthermore some changes must be made to ensure consistent
x/y labeling
* Sometimes drawing produces extremely weird output, perhaps this is
related to odd plot points?
* Fast/Unknown drawing mode are confusing. Fast is Fast for bpp8 and
medium otherwise. Unknown is Medium for bpp8 and Fast otherwise, at
least observations seem to suggest this.