[PAUSED, SLOW] Rework command to use a GADT
Previously the command module used an unsafe procedure: Send command and
then send arguments. Using ADTs we can encode the arguments along with
the command and, with GADTs, we can also specify the return value.
Additionaly it is stated only that GADTs are benificial to performance
as they allow the compiler to know more about code paths.
Implementing GADTs proved to be relatively easy, we now have separated
commands for external use and interal use. The former exposes the
functionality and the latter represents the actual commands sent.

Unfortunately this change resulted (without further investigation) in a
considerable slowdown of about 4x. Cause is unknown at this time and it
was tested without flambda (not enough RAM for compilation). This
approach would be extremely beneficial for simplicity of use but this
performance loss is inacceptable.

This commit furthermore represents the last commit for the time being as
the new semester at university started. I will most likely continue in august.